This is a first partial attempt at a specification of the buffer protocols.
Having a good specification for those protocols will allow us to
substitute better implementations, for instance as soon as the
flexichain library becomes available.

At the moment, I only wrote the simple ones like `clef' and `staff'.
It will get messier for major container classes like `bar', `segment',
etc.  In particular, how cursors are managed needs to be specified in
detail.

Now might also be the time to question the wisdom of having notes be
read-only, something that complicated the code for ties as indicated
by Christophe Rhodes.