Updated the `getting started' to reflect the new and altered commands for staves and layers.
Updated the command summary in the reference manual for the same reason.

Added items to the release notes to reflect user-visible changes.

Removed and modified some items in the `future plans' section to
reflect some improvements already made.

At some point in the future, I need to update the `internals' part to
reflect the existence of lyrics, etc., but since the way it is
implemented might change some, there is no point in doing that right now.