Gsharp definitely has lyrics.
When a lyrics layer is selected, the ordinary alphabetic keys append
characters to the preceding lyrics element.  use C-Space to insert a
new lyrics element.  Use C-h to erase the last character of a lyrics
element.  It's a bit clunky, but it works.  I seriously doubt the full
power of Goatee is needed to edit lyrics.

I noticed that (probably unintentionally) McCLIM can handle Unicode
texts in the form of vectors of integers to draw-text.  It probably
just passes them on to the port-specific functions, and it happens to
work with clx.  It probably won't work on all backends, but it would
be nice if it did.  The default font on my machine can probably only
do character codes up to 255, but with a good font, it should be able
to do up to 65535.  But right now, I can get the German national
characters, which are needed for Bach.

Check out the new file Scores/bach181-lyrics.gsh.

In it, you will also notice that there is one place in which words
overlap.  To fix that (which is not just a lyrics problem, but also
happens with clusters with many accidentals) Gsharp must learn to
compute how much extra room an element might need, and adjust spacing
accordingly,

Unfortunately McCLIM does not handle input from a US-international
keyboard, so I actually had to put in the ü's manually, but it should
work for others.

With lyrics, MIDI file generation is most likely broken, but I haven't
checked that yet. It should just skip lyrics layers.