Lyrics input now allows for accents, provided that your keyboard is capable of generating dead keys for these accents.  It only works for X11 at the moment, I would think.
The way I do it is not great, since now I get messages about missing
commands for the control keys.  I have to think of a better way of
doing it.

Loading a file now calls select-layer so that the appropriate mode is
selected.  Loading the bach score with lyrics now immediately lets
you edit the lyrics.