I am proud to announce the first reasonably complete version of the `flexichain' library.  The bad news is that the API does not quite correspond to what was previously announced.  There are three pieces of good news, however: First the API is better now, second, the documentation has been updated to be consistent with the modified API, and third the code is much simpler than it would have been with the original API.
There are still some inconsistencies between the documentation and the
code (names of conditions, missing pieces of the API in the doc), but
it is mostly there.

Oh, and since it was so hard to test, I wrote a CLIM application to
test it.  The upper half is a black-box test and the lower half is
the white-box test.  Let me know what you think.

Before integrating the Flexichain library into Gsharp, I would like to
test it some more, so if anyone feels like doing that and reporting
bugs, that would be fantastic.