Completely modified the implementation of cursors.
Now, a cursorchain can hold a large number of cursors without any
negative impact on performance.

For cursorchains, the flexichain buffer now has a parallel that holds
per-element lists of cursors that stick to that element.

Introduced new generic functions in the internal protocol fill-gap and
resize-buffer.

Updated the documentation accordingly.