Better filename completer that works for both SBCL and CMUCL.
Gsharp no longer destroys the port before starting up.  This is in
preparation to run applications from the listener, or from some other
application.

Made some cosmetic changes after having learned about the existence of
CONSTANTLY.

*gsharp-frame* (is it still needed?) is no longer setf'ed but bound,
so that each thread has its own copy.

Added numbering of the segments of a buffer as :after method on
initialize-instance on a buffer.