Removed package.lisp,  I think I might start naming the package files mumble-package.lisp for the module mumble.lisp.
Added flexichain-package.lisp to replace package.lisp.

Added a new module for managing cursors: skiplist.  This is a
rotatable skiplist, which is like a skiplist, except that it allows
"rotations", i.e. a prefix interval of keys can be moved to the end,
and q suffix interval of keys can be moved to the beginning.  All this
in time proportional to log(n) + m (probabilistically, not
worst-caase), where n is the size of the skiplist and m is the number
of keys that need to be moved.

I have (somewhat) tested the skiplist module, and it appears to work,
though I may have broken something during the last minor update.

The file flexicursor.lisp has been updated to use the skiplist module.
I have not yet tested the result, though.  Consider this commit as a
backup as opposed to a commit of a version believed to be stable.