Changed the sheet transformation of the score pane to be the default for CLIM stream panes (0,0) in the upper-left corner.  This was in order to simplify the rest of the code, and in particular the output recording stuff.  There are probably some edge cases that don't yet work like they are supposed to with the possibility of off-by-a-pixel errors.
Removed all the rectangle output records in favor of calls to
draw-rectangle*.

Temporarily removed the double buffering as a preparation for better
seeing what is going on with incremental redisplay.

Getting incremental redisplay to work might require fixing a problem
in McCLIM which does not necessarily rely on the output-record
protocol, but instead sometimes assume the existence of a slot in the
record, whereas no such slot is required by the specification.  The
result is that the user cannot define his or her own output records
and have them work with incremental redisplay.