Prepared Gsharp for multi-buffer, multi-frame, and multi-view features.
This modification involved getting rid of the frame-global `buffer'
and `cursor' slots.  Now, a new class `gsharp-pane', a subclass of
score-pane, contains a slot for a view.  The idea is that a pane has a
particular view on display, and the view contains the buffer and the
cursor to be displayed in the pane.

Eventually C-x b will be used to change the view on display in the
current pane, C-x k will kill the view (and if it is the last view
that displays a certain modified buffer, the user will be asked to
confirm), C-x 2 will clone the view into a new top-level window.

There will also be commands to alter the class of the current view to
obtain parts views etc.  At least, this corresponds to my current
thinking.

This modification was obtained without using Emacs (except for typing
this message).  Instead I used the CLIM Desktop.  Specifically, I used
Climacs for editing source code with Swine for incremental compilation
and calling Closure to read CLHS documentation, and the CLIM Listener
to compile and execute Gsharp.  While Climacs and the other tools
still have some quirks, I must say I am VERY impressed with what they
can already do.