The default key signature of a staff is now represented by an instance of the new class `key-signature', rather than by just a vector.  The commands `com-more-sharps' and `com-more-flats' now call new protocol generic functions on the key signature.
I used the suggestion from the patch by Christophe Rhodes to introduce
a new class `rhythmic-element' below `element' and move slots that
have to do with duration to that new class (rbeams, lbeams, dots).
The `key-signature' class does not inherit from `rhythmic-element',
but instead directly from `element'.

In order to avoid having to alter the external format yet again, the
reader tests whether a vector was read as the key signature, and if
so, replaces it by an instance of the new class.

As a nice side effect, I was able to remove the symbol
`invalidate-everything-using-staff' from the list of exported symbols
of `measure.lisp', because it is now used by the :after methods on
`more-sharps' and `more-flats', defined in the same package.

What I haven't done (I'll let Christophe do it, unless he takes too
long) is to incorporate the parts from Christophe's patch that make it
possible to insert key signatures as elements into layers.