Added a new package and a new file ESA-BUFFER allowing buffers to be named, to be associated with a file name, and to have a `needs-saving' and a `read-only' flag.
Added a new package and a new file ESA-IO containing
application-independent functionality to create buffers from files,
and to save buffers to files.  This package also supplies filename
completion.  Most of the code was adapted from Climacs.

Abstracted out all Gsharp-specific I/O to ESA-IO.  In particular, this
means that we now have commands such as C-x C-s, and C-x C-w, which we
didn't before.

The old I/O code is still there.  Cleanup is next.