The new font-rendering code is now in there, but is not yet being used. The reason for that is that I still have not managed to get output recording for designs right.
Once that problem is fixed, I am planning to gradually move to the new
system, debugging the glyphs one at a time.  The code for the glyphs
has been tested in a separate context, but there might be unforeseen
problems.

The new system allows designs to be drawn in any color and
transformation by being rendered to anti-aliased pixmaps, so there is
no need to use special gray pixmaps.  It might be worthwhile thinking
about moving beam drawing to this new system one day.

This font rendering system should be easy for PostScript output, but
I haven't attacked that problem yet.