Half of a new font viewer application.
The half that exists is that it shows the resulting anti-aliased
design, with or without a reference staff.

What remains to write is showing the pixel view with grids and
reference points, and perhaps also end points and reference points of
the Bezier segments.  This should be a relatively simple matter of
calling render-to-array and showing the result.