Moved things around a bit inside bezier.lisp to make it easier to render to an array from the font viewer.
Implemented pixel viewing in the font viewer.  Initial results
indicate that the G-clef is pretty good, and there must be something
wrong either with the way the pixmap gets generated or with the way it
gets copied to the pane (it is too far down).  Initial result also
indicate that the C-clef is completely wrong and incredibly ugly at
sizes above 6.