Fixed a bug in bezier.lisp that made the resulting pixmap positioned in the wrong place sometimes.
Fixed a bug in sdl.lisp that computed the wrong xoffset (in particular
for noteheads).

Fixed the whole notehead so that its vertical position is not offset
as it should.

Fixed the C clef so that it no longer goes below the bottom line.

Improved the font viewer by having the bounding box of the glyph drawn
in :pixel mode.