Dots!
Specifically, augmentation dots.  Get their x- and y- positions more
right, which sometimes entails not drawing a dot at all, sometimes
adjusting the position for a dot downwards, and (when a flag is drawn or
there is a suspended note in a flag-up situation) involves shifting the
entire column of dots rightwards.

Add an example score full of things we got wrong.