Somewhat hacky motion-by-layout commands (line- and page-based).  These was by far the most annoying motion commands to be lacking when actually editing scores, in my opinion; C-a and C-e are absolutely wired into my fingers.
Future candidates for inclusion: motion by layer (ideally in a way
that's related to the layout, so that in the common case where there is
one layer per staff the Right Thing happens); motion by
note-in-associated-layer (lyrics).

Also remove C-h binding for delete-current-element, and put Backspace in
instead; this lets the ESA help bindings work.