Alright, let's try this: more correct key signatures, I hope.
The major change to the protocol is that REMOVE-ELEMENT takes as a
required argument the bar as well as the element; this allows more
symmetric methods to be written for the various stealth mixin bits of
functionality.

Key signatures are elements, as before, within a layer.  However, they
are also kept on a list sorted by sequence in a slot of the staff, and
KEYSIG is responsible for checking the relevant staff for other key
signature elements.  Editing actions or commands are also responsible
for maintaining this list sorted in the right order.  New almost-correct
function for testing the temporal-and-logical ordering of elements.

Drawing code now computes the correct key signature for the each staff;
linebreaking is done with a conservative assumption for how wide the key
signature will be.

Please test.