Better tie drawing:
Tie direction is taken from the final stem direction of the first note.
This is not actually right, but it's incrementally better than ignoring
the stem directions completely.

Draw a stub tie forward if no matching note is found.  (This needs to
happen for unpaired backwards ties too, but the way DRAW-TIES is
currently structured makes that mildly tricky.)