Fixed the first problem reported by Taylor Campbell, namely that the locations of certain notes compared to the stem are not computed correctly.
The main problem was that compute-element-parameters was called only
for elements that have been modified.  However, those parameters might
change as a result of beaming with other elements, so we need to call
compute-element-parameters on all elements.

While searching for the cause of this problem, I also noticed that a
stem direction that was forced to change as a result of beaming with a
subsequent element (such that the stem direction for the entire beam
group was different from the nominal one for the cluster) did not
return to its nominal stem direction when the subsequent element was
deleted.  The reason for this problem was that
compute-final-stem-direction was not called unless any of the elements
of the beam group was marked as modified.  When the subsequent element
was removed, none of the remaining elements was marked that way.  The
fix was to always call compute-final-stem-direction.

I also discovered another problem from Taylor's example, namely that
sometimes, the stem direction of a beam group in which each cluster
has the same nominal stem direction can be different from the nominal
stem direction of the individual clusters.  The reason for this
problem seems to be that the algorithm for computing the stem
direction of a beam group is just wrong.  I am still working on this
problem.